+++ tests/unit/test_cloud_shelf.py
@@ -11,7 +11,7 @@ def test_factory_usage():
     """
     Ensure the usage of the factory.
     """
-    filename = "./tests/configurations/azure.ini"
+    filename = "does_not_exists.ini"
     provider = "azure-blob"
     config = {42: 42}
     flag = "c"
@@ -34,11 +34,36 @@ def test_factory_usage():
         assert isinstance(cs.dict.db, Mock)
 
 
+def test_loader_usage():
+    """
+    Ensure the usage of the loader.
+    """
+    filename = "does_not_exists.ini"
+    provider = "azure-blob"
+    config = {42: 42}
+    flag = "c"
+    protocol = pickle.HIGHEST_PROTOCOL
+    writeback = False
+
+    factory = Mock()
+    loader = Mock()
+    cloud_database = Mock()
+
+    loader.return_value = provider, config
+    factory.return_value = cloud_database
+    cloud_database.exists.return_value = False
+
+    with CloudShelf(
+        filename, flag, protocol, writeback, loader=loader, factory=factory
+    ) as cs:
+        loader.assert_called_once_with(filename)
+
+
 def test_provider_configuration():
     """
     Ensure the provider returns by the factory is configured.
     """
-    filename = "./tests/configurations/azure.ini"
+    filename = "does_not_exists.ini"
     provider = "azure-blob"
     config = {42: 42}
     flag = "c"
